refactor: edit component imports
