--- src/components/Button/index.tsx
@@ -1,2 +1,0 @@
-import Button from './Button';
-export default Button;
