--- src/components/Cards/index.tsx
@@ -1,2 +1,0 @@
-import Cards from './Cards';
-export default Cards;
