+++ src/components/Pagination.tsx
@@ -1,0 +1,3 @@
+export default function Pagination () {
+    return <></>
+}
