+++ src/components/Results.tsx
@@ -1,8 +1,8 @@
-import { SearchContext } from '../../modules/context';
-import { searchData } from '../../services/api.service';
-import Cards from '../Cards';
-import { Shows } from '../../modules/types';
+import { SearchContext } from '../modules/context';
+import { searchData } from '../services/api.service';
+import { Shows } from '../modules/types';
 import { useContext, useState, useEffect } from 'react';
+import { Pagination, Cards } from './components';
 
 function Results() {
   const context = useContext(SearchContext);
@@ -29,17 +29,20 @@ function Results() {
 
   return (
     <>
-      {isLoading ? (
-        <div className="text-white py-4">Loading...</div>
-      ) : (
-        <article className="text-white py-4">
-          {data && data.length !== 0 ? (
-            <Cards data={data} />
-          ) : (
-            'nothing to show'
-          )}
-        </article>
-      )}
+      <Pagination></Pagination>
+      <section>
+        {isLoading ? (
+          <div className="text-white py-4">Loading...</div>
+        ) : (
+          <article className="text-white py-4">
+            {data && data.length !== 0 ? (
+              <Cards data={data} />
+            ) : (
+              'nothing to show'
+            )}
+          </article>
+        )}
+      </section>
     </>
   );
 }
