--- src/components/Results/index.tsx
@@ -1,3 +1,0 @@
-import Results from './Results';
-
-export default Results;
