+++ src/components/ResultsCard.tsx
@@ -1,4 +1,4 @@
-import { ShowsProps } from '../../modules/interfaces';
+import { ShowsProps } from '../modules/interfaces';
 
 type ResultCardProps = {
   item: ShowsProps;
@@ -6,7 +6,7 @@ type ResultCardProps = {
 
 export default function ResultsCard(props: ResultCardProps) {
   return (
-    <article className="bg-slate-500 max-w-xs p-4 rounded">
+    <div className="bg-slate-500 max-w-xs p-4 rounded">
       <a
         href={typeof props.item.imdbUrl === 'string' ? props.item.imdbUrl : ''}
         className="underline my-1 hover:text-orange-300 transition"
@@ -22,6 +22,6 @@ export default function ResultsCard(props: ResultCardProps) {
         <span className="">IMDB Rating: </span>
         <span className="font-bold">{props.item.imdbRating}</span>
       </div>
-    </article>
+    </div>
   );
 }
