--- src/components/ResultsCard/index.tsx
@@ -1,2 +1,0 @@
-import ResultsCard from './ResultsCard';
-export default ResultsCard;
