--- src/components/SearchInput/index.tsx
@@ -1,2 +1,0 @@
-import SearchInput from './SearchInput';
-export default SearchInput;
