+++ src/components/components.tsx
@@ -1,0 +1,8 @@
+import Pagination from './Pagination';
+import Cards from './Cards';
+import Search from './Search';
+import Results from './Results';
+import SearchInput from './SearchInput';
+import ResultsCard from './ResultsCard';
+import Button from './Button';
+export { Pagination, Cards, Button, Search, SearchInput, Results, ResultsCard };
